Had to move the dummy data out of public
